added image option

+++ app/Http/Controllers/PostsController.php
@@ -3,12 +3,24 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 
 class PostsController extends Controller
 {
+
+
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -46,13 +58,28 @@ class PostsController extends Controller
 
         $this->validate($request,[
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
         //
+
+        if($request->hasFile('cover_image')){
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $fileNameToStore = $fileName .'_'.time().'.'.$extension;
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        }
+        else{
+            $fileNameToStore = 'noimage.jpeg';
+        }
 
       $post = new Post;
       $post->title = $request->input('title');
       $post->body = $request->input('body');
+      $post->user_id = auth()->user()->id;
+      $post->cover_image = $fileNameToStore;
       $post->save();
 
       return redirect('/posts')->with('success', 'Post created.');
@@ -83,6 +110,9 @@ class PostsController extends Controller
     {
 
         $post =  Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
 
         return view('posts.edit')->with('post', $post);
         //
@@ -100,13 +130,24 @@ class PostsController extends Controller
         //
         $this->validate($request,[
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
         //
+        if($request->hasFile('cover_image')){
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $fileNameToStore = $fileName .'_'.time().'.'.$extension;
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
 
+        }
       $post = Post::find($id);
       $post->title = $request->input('title');
       $post->body = $request->input('body');
+      if($request->hasFile('cover_image')){
+          $post->cover_image = $fileNameToStore;
+      };
       $post->save();
 
       return redirect('/posts')->with('success', 'Post updated.');
@@ -121,7 +162,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         //
+
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
+        if($post->cover_image !== 'noimage.jpg'){
+            Storage::delete('public/cover_images'.$post->cover_image);
+        }
+
+        // $post = Post::find($id);
         $post->delete();
         return redirect('/posts')->with('success', 'Post removed.');
